requirements.txt
streamlit>=1.20
rank_bm25>=0.2.2
nltk>=3.8
numpy>=1.24
langchain==0.3.27
langchain_groq==0.3.0
langsmith
python-dotenv>=1.0.1
